+++ .eslintrc.js
@@ -1,0 +1,48 @@
+module.exports = {
+	'parser': 'babel-eslint',
+	'env': {
+		'browser': true,
+		'es6': true
+	},
+	'extends': [
+		'eslint:recommended',
+		"plugin:react/recommended"
+	],
+	'globals': {
+		'Atomics': 'readonly',
+		'SharedArrayBuffer': 'readonly'
+	},
+	'parserOptions': {
+		'ecmaFeatures': {
+			'jsx': true
+		},
+		'ecmaVersion': 2018,
+		'sourceType': 'module'
+	},
+	'plugins': [
+		'react'
+	],
+	'rules': {
+		'indent': [
+			'error',
+			'tab'
+		],
+		'linebreak-style': [
+			'error',
+			'unix'
+		],
+		'quotes': [
+			'error',
+			'single'
+		],
+		'semi': [
+			'error',
+			'never'
+		]
+	},
+	'settings': {
+		'react': {
+			'version': 'detect'
+		}
+	}
+}
