+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "editor.insertSpaces": false,
+	"editor.detectIndentation": false
+}
