+++ src/Components/AppContainer.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export class AppContainer extends React.Component {
+	constructor(props) {
+		super(props)
+	}
+
+	render() {
+		return <div>
+			
+		</div>
+	}
+}
